test: host native inspectability lifecycle coverage
