+++ Tests/WebInspectorProxyKitTests/NativeInspectablePageTests.swift
@@ -1,47 +1,6 @@
 #if canImport(UIKit)
 import Testing
-import UIKit
-import WebKit
 @testable import WebInspectorProxyKit
-
-@MainActor
-@Test
-func nativeInspectablePageRestoresOriginalInspectability() {
-    let webView = WKWebView(frame: .zero)
-    webView.isInspectable = false
-
-    let page = NativeInspectablePage(webView: webView)
-
-    #expect(webView.isInspectable == true)
-
-    page.restoreInspectabilityIfNeeded()
-
-    #expect(webView.isInspectable == false)
-
-    page.restoreInspectabilityIfNeeded()
-
-    #expect(webView.isInspectable == false)
-}
-
-@MainActor
-@Test
-func overlappingNativeInspectablePagesRestoreOnlyAfterLastOwner() {
-    let webView = WKWebView(frame: .zero)
-    webView.isInspectable = false
-
-    let firstPage = NativeInspectablePage(webView: webView)
-    let secondPage = NativeInspectablePage(webView: webView)
-
-    #expect(webView.isInspectable == true)
-
-    secondPage.restoreInspectabilityIfNeeded()
-
-    #expect(webView.isInspectable == true)
-
-    firstPage.restoreInspectabilityIfNeeded()
-
-    #expect(webView.isInspectable == false)
-}
 
 @MainActor
 @Test
