+++ WebInspectorKit.xcworkspace/xcshareddata/xcschemes/WebInspectorKitTests.xcscheme
@@ -123,7 +123,13 @@
             </BuildableReference>
             <SelectedTests>
                <Test
+                  Identifier = "WebInspectorProxyKitIntegrationTests/testNativeInspectablePageRestoresOriginalInspectability()">
+               </Test>
+               <Test
                   Identifier = "WebInspectorProxyKitIntegrationTests/testNativeInspectorReconnectsAfterWebContentProcessTermination()">
+               </Test>
+               <Test
+                  Identifier = "WebInspectorProxyKitIntegrationTests/testOverlappingNativeInspectablePagesRestoreOnlyAfterLastOwner()">
                </Test>
                <Test
                   Identifier = "WebInspectorProxyKitIntegrationTests/testSecondNativeInspectorAttachmentFailsWithoutReplacingDelegate()">
